requirements.txt
amqp==5.3.1
arrow==1.3.0
asgiref==3.8.1
billiard==4.2.1
click==8.2.1
colorama==0.4.6
Django==5.2.3
django-cors-headers==4.7.0
django-debug-toolbar==5.2.0
django-filter==25.1
django-redis==6.0.0
django-rq==3.0.1
djangorestframework==3.16.0
pillow==11.2.1
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
redis==6.2.0
rq==2.4.0
sqlparse==0.5.3
times==0.7
types-python-dateutil==2.9.0.20241206
tzdata==2025.2
whitenoise==6.9.0